Permissons

+++ StudeasySystem/ejbModule/de/studeasy/schedulemanager/StudeasyScheduleService.java
@@ -30,7 +30,11 @@ import de.studeasy.systeminterfaces.StudeasyException;
 import de.studeasy.util.DtoAssembler;
 
 /**
- * 
+ * Der StudeasyScheduleService stellt eine Statless Session Bean dar,
+ * die den eigentlichen Kern des Server-Systems darstellt.
+ * Hier ist die Hauptlogik implementiert und somit greift der Client
+ * auf die Methoden dieses Services zu, um an seine gewünschten Daten zu kommen.
+ * TODO
  * @author Tobias Riegel & Andreas Prischep
  *
  */
@@ -96,29 +100,43 @@ public class StudeasyScheduleService implements IStudeasyScheduleService {
 	
 	
 	//---------------------------CREATE HOMEWORK------------------------------------------------------------
-	//TODO Muss noch richtig implementiert werden.
 	@Override
 	public BooleanResponse createHomework(int sessionID, int lessonID, String description)  {
 		
 		BooleanResponse response = new BooleanResponse();
+		IPerson user = dao.findPersonByID(dao.findSessionByID(sessionID).getUserID());
 		
-	//TODO mit sessionID noch auf Berechtigung prüfen
+		if(user instanceof ITeacher) {
 			ILesson lesson = dao.findLessonByID(lessonID);
 			IHomework homework = new Homework(description, lesson);
 			lesson.addNewHomework(homework);
 			response.setSuccessfull(true);
+		}
+		else {
+			response.setSuccessfull(false);
+			response.setMessage("Der angemeldete User ist keine Lehrer und darf somit keine Hausaufgaben anlegen.");
+			//TODO Neuen Fehlercode definieren?
+			//response.setReturnCode(returnCode);
+		}
 		return response;
 	}
 	//---------------------------REMOVE HOMEWORK------------------------------------------------------------
 	@Override
 	public BooleanResponse removeHomework(int sessionID, int homeworkID)  {
+		BooleanResponse response = new BooleanResponse();
+		boolean successfull = false;
+		IPerson user = dao.findPersonByID(dao.findSessionByID(sessionID).getUserID());
 		
-			boolean successfull = dao.removeHomeworkByID(homeworkID);
+		if(user instanceof ITeacher) {
+			successfull = dao.removeHomeworkByID(homeworkID);
+		}
+		else {
+			response.setMessage("Der angemeldete User ist keine Lehrer und darf somit keine Hausaufgaben anlegen.");
+			//TODO Neuen Fehlercode definieren?
+			//response.setReturnCode(returnCode);
+		}
 			
-			BooleanResponse response = new BooleanResponse();
-			
-			response.setSuccessfull(successfull);
-		
+		response.setSuccessfull(successfull);
 		return response;
 	}
 	
@@ -127,7 +145,6 @@ public class StudeasyScheduleService implements IStudeasyScheduleService {
 	 * Gibt eine Liste der Unterrichtsstunden an einem bestimmten Tag
 	 * für eine bestimmte Person zurück.
 	 * Die Liste ist leer, wenn an dem Tag kein für die Person kein Unterricht ist.
-	 * Wenn null zurückgeben wird, waren die Parameter falsch.
 	 */
 	
 	@Override
@@ -141,6 +158,7 @@ public class StudeasyScheduleService implements IStudeasyScheduleService {
 		try {
 			StudeasySession session = getSession(sessionID);
 			IPerson person = dao.findPersonByID(session.getUserID());
+			
 			if(person instanceof IPupil) {
 				IPupil pupil = (IPupil) person;
 				lessons = pupil.getCourse().getLessons();	
@@ -189,8 +207,7 @@ public class StudeasyScheduleService implements IStudeasyScheduleService {
 	/** 
 	 * Gibt eine Liste der Unterrichtsstunden in einem bestimmten Zeitraum
 	 * für eine bestimmte Schulklasse in einem bestimmten Fach zurück.
-	 * Die Liste ist leer, wenn in dem Zeitruam kein für die Schulklasse in dem Fach kein Unterricht ist.
-	 * Wenn null zurückgeben wird, waren die Parameter falsch.
+	 * Die Liste ist leer, wenn in dem Zeitraum für die Schulklasse in dem Fach kein Unterricht ist.
 	 * Das startDate und endDate sind inklusive. Also Unterrichtsstunden die am 
 	 * start- oder endDate stattfinden befinden sich in der zurückgegebenen Liste.
 	 */
@@ -237,7 +254,7 @@ public class StudeasyScheduleService implements IStudeasyScheduleService {
 	/**
 	 * Gibt eine Liste von Hausaufgaben für einen bestimmten Schüler zu einem bestimmten Zeitraum zurück.
 	 * Die Liste ist leer, wenn dieser Schüler zu diesem Zeitraum keine Hausaufgaben hat.
-	 * Es wird null zurückgegeben, wenn die personID nicht zu einem Schüler gehört.
+	 * Es Liste ist null, wenn die personID nicht zu einem Schüler gehört.
 	 */
 	@Override
 	public HomeworkListResponse getHomeworksForPupil(int sessionID, Date startDate,
@@ -265,13 +282,16 @@ public class StudeasyScheduleService implements IStudeasyScheduleService {
 				
 				response.setHomeworkList(dto.makeHomeworkDTO(homeworks));
 				}
-			else
-				response.setHomeworkList( null);
-			
+			else {
+				response.setHomeworkList(null);
+				response.setMessage("Der angemeldete User ist keine Schüler.");
+				//TODO Neuen Fehlercode definieren?
+				//response.setReturnCode(returnCode);
+			}
 			
 		}
 		catch (StudeasyException e) {
-			response.setHomeworkList( null);
+			response.setHomeworkList(null);
 			response.setReturnCode(e.getErrorCode());
 			response.setMessage(e.getMessage());
 		}
